physics modules are moved to physics dir; fixed enemies pushing out of the level;
