+++ src/engine.cpp
@@ -27,6 +27,7 @@ extern "C" {
 #include "render/camera.h"
 #include "render/render.h"
 #include "script/script.h"
+#include "physics/physics.h"
 #include "gui/gui.h"
 #include "vt/vt_level.h"
 #include "game.h"
@@ -39,7 +40,6 @@ extern "C" {
 #include "world.h"
 #include "resource.h"
 #include "engine.h"
-#include "physics.h"
 #include "controls.h"
 #include "trigger.h"
 #include "character_controller.h"
