+++ src/game_camera.cpp
@@ -6,11 +6,11 @@
 #include "core/obb.h"
 #include "core/system.h"
 #include "render/camera.h"
+#include "physics/physics.h"
 #include "engine.h"
 #include "controls.h"
 #include "room.h"
 #include "world.h"
-#include "physics.h"
 #include "skeletal_model.h"
 #include "entity.h"
 #include "character_controller.h"
