+++ src/render/render.cpp
@@ -12,6 +12,7 @@
 #include "../core/polygon.h"
 #include "../core/obb.h"
 #include "../script/script.h"
+#include "../physics/physics.h"
 #include "../vt/tr_versions.h"
 #include "camera.h"
 #include "render.h"
@@ -26,7 +27,6 @@
 #include "../entity.h"
 #include "../character_controller.h"
 #include "../engine.h"
-#include "../physics.h"
 
 CRender renderer;
 
@@ -824,7 +824,7 @@ void CRender::DrawSkeletalModel(const lit_shader_description *shader, struct ss_
         {
             Mat4_Mat4_mul(mvTransform, mvMatrix, btag->full_transform);
             qglUniformMatrix4fvARB(shader->model_view, 1, false, mvTransform);
-            
+
             Mat4_Mat4_mul(mvpTransform, mvpMatrix, btag->full_transform);
             qglUniformMatrix4fvARB(shader->model_view_projection, 1, false, mvpTransform);
