+++ src/room.cpp
@@ -10,8 +10,8 @@
 #include "core/polygon.h"
 #include "core/obb.h"
 #include "render/frustum.h"
+#include "physics/physics.h"
 #include "engine.h"
-#include "physics.h"
 #include "entity.h"
 #include "mesh.h"
 #include "trigger.h"
