+++ src/script/script_world.cpp
@@ -19,6 +19,7 @@ extern "C" {
 #include "../render/camera.h"
 #include "../render/render.h"
 #include "../gui/gui.h"
+#include "../physics/physics.h"
 #include "../vt/tr_versions.h"
 #include "../mesh.h"
 #include "../skeletal_model.h"
@@ -28,7 +29,6 @@ extern "C" {
 #include "../entity.h"
 #include "../world.h"
 #include "../engine.h"
-#include "../physics.h"
 #include "../controls.h"
 #include "../game.h"
 #include "../gameflow.h"
