+++ src/world.cpp
@@ -23,6 +23,7 @@ extern "C" {
 #include "render/bordered_texture_atlas.h"
 #include "script/script.h"
 #include "state_control/state_control.h"
+#include "physics/physics.h"
 #include "gui/gui.h"
 #include "vt/vt_level.h"
 #include "audio.h"
@@ -33,7 +34,6 @@ extern "C" {
 #include "entity.h"
 #include "character_controller.h"
 #include "engine.h"
-#include "physics.h"
 #include "gameflow.h"
 #include "resource.h"
 #include "inventory.h"
